backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
pydantic==2.7.3
python-multipart==0.0.9
pandas==2.2.2
scikit-learn==1.5.0
prophet==1.1.5
httpx==0.27.0
huggingface-hub==0.23.0
transformers==4.41.2
ollama==0.1.9
prometheus-client==0.20.0
structlog==24.1.0
pytest==8.3.2
python-dotenv==1.0.1
orjson==3.10.3
numpy==1.26.4
